Check the stream resource is valid before closing
When disconnecting a StreamConnection sometimes PHP will trigger an
error because the resource is not a valid stream resource. Presumably
the stream has already been closed.

It's been addressed for now by checking if the resource is valid before
trying to close it.

Presumably this, or other errors are known upstream as fwrite and other
calls are error silenced and will retry on error.